pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["malakkhasawneh2 <malakkhasawneh2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
black = "^21.12b0"
flake8 = "^4.0.1"
djangorestframework = "^3.13.0"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
